requirements.txt
pytest
allure-pytest
pytest-html
requests
selenium
selenium-page-factory